optimized the paths
