--- src/Components/Navbar/Navbar.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-const Navbar = () => {
-  return (
-    <div>
-      <h2 className='text-left bg-center font-sans '>Hello World</h2>
-      <h2 className='font-sans text-rose-600 text-2xl align-text-bottom'>I am Kaustubh</h2>
-    </div>
-  )
-}
-
-export default Navbar
